@@ -26,8 +26,8 @@
 
     <properties>
         <core.version>2.0.0-SNAPSHOT</core.version>
-        <spring-boot.version>2.1.6.RELEASE</spring-boot.version>
-        <spring-cloud.version>Greenwich.SR1</spring-cloud.version>
+        <spring-boot.version>2.4.1</spring-boot.version>
+        <spring-cloud.version>2020.0.0</spring-cloud.version>
     </properties>
 
     <dependencyManagement>
