@@ -2,8 +2,8 @@ package com.dch.core.common.util;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import org.junit.BeforeClass;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
 
 import javax.validation.ConstraintViolationException;
 import javax.validation.Validation;
@@ -12,6 +12,7 @@ import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Positive;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 /**
  * Test {@link ValidationUtils}.
@@ -22,22 +23,29 @@ import static org.assertj.core.api.Assertions.assertThatCode;
  */
 public class ValidationUtilsTest {
 
-    @BeforeClass
+    @BeforeAll
     public static void setUpOnce() {
         ValidatorFactory validatorFactory = Validation.buildDefaultValidatorFactory();
         ValidationUtils.setValidator(validatorFactory.getValidator());
     }
 
     @Test
-    public void validate_validModel_noErrorThrown() {
+    public void testValidate_withValidModel_thenExpectNoErrorThrown() {
         Model model = new Model("ten", 10);
         assertThatCode(() -> ValidationUtils.validate(model)).doesNotThrowAnyException();
     }
 
-    @Test(expected = ConstraintViolationException.class)
-    public void validate_invalidModel_throwConstraintViolationException() {
+    @Test
+    public void testValidate_withNewValidator_thenExpectValidatorIsIgnored() {
+        Model model = new Model("ten", 10);
+        ValidationUtils.setValidator(Validation.buildDefaultValidatorFactory().getValidator());
+        assertThatCode(() -> ValidationUtils.validate(model)).doesNotThrowAnyException();
+    }
+
+    @Test
+    public void testValidate_withInvalidModel_throwConstraintViolationException() {
         Model model = new Model(" ", -1);
-        ValidationUtils.validate(model);
+        assertThrows(ConstraintViolationException.class, () -> ValidationUtils.validate(model));
     }
 
     @Getter
