@@ -22,6 +22,7 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.stream.Collectors;
 
+import static net.logstash.logback.argument.StructuredArguments.kv;
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 
 /**
@@ -29,7 +30,6 @@ import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
  *
  * @author david.christianto
  * @version 2.0.0
- * @see BaseEndpoint
  */
 @RestControllerAdvice
 public class ExceptionHandlerEndpoint extends BaseEndpoint {
@@ -48,11 +48,12 @@ public class ExceptionHandlerEndpoint extends BaseEndpoint {
      * @return {@link ResponseDto Response body} of message not readable.
      */
     @ApiResponse(responseCode = "400", description = "HTTP message not readable because of malformed JSON request",
-            content = @Content(mediaType = APPLICATION_JSON_VALUE, schema = @Schema(implementation = ResponseDto.class)))
+            content = @Content(mediaType = APPLICATION_JSON_VALUE, schema = @Schema(implementation =
+                    ResponseDto.class)))
     @ExceptionHandler(HttpMessageNotReadableException.class)
     @ResponseStatus(HttpStatus.BAD_REQUEST)
     public ResponseDto<Void> handleHttpMessageNotReadableException(HttpMessageNotReadableException ex) {
-        logger.error("HTTP message error! " + GeneralStatus.MESSAGE_NOT_READABLE, ex);
+        logger.error("HTTP message doesn't readable! {}", kv("status", GeneralStatus.MESSAGE_NOT_READABLE), ex);
         return getResponseBuilder(Void.class)
                 .setGeneralStatus(GeneralStatus.MESSAGE_NOT_READABLE)
                 .build();
@@ -64,12 +65,14 @@ public class ExceptionHandlerEndpoint extends BaseEndpoint {
      * @param ex the {@link MethodArgumentNotValidException}
      * @return {@link ResponseDto Response body} of validation error.
      */
-    @ApiResponse(responseCode = "400", description = "Method argument not valid because of failed to validate JSON request",
-            content = @Content(mediaType = APPLICATION_JSON_VALUE, schema = @Schema(implementation = ResponseDto.class)))
+    @ApiResponse(responseCode = "400", description = "Method argument not valid because of failed to validate JSON " +
+            "request",
+            content = @Content(mediaType = APPLICATION_JSON_VALUE, schema = @Schema(implementation =
+                    ResponseDto.class)))
     @ExceptionHandler(MethodArgumentNotValidException.class)
     @ResponseStatus(HttpStatus.BAD_REQUEST)
     public ResponseDto<Map> handleMethodArgumentNotValidException(MethodArgumentNotValidException ex) {
-        logger.error("HTTP message error! " + GeneralStatus.ARGUMENT_NOT_VALID, ex);
+        logger.error("Method arguments don't valid! {}", kv("status", GeneralStatus.ARGUMENT_NOT_VALID), ex);
 
         Map<String, String> errorsMap = ex.getBindingResult().getAllErrors().stream()
                 .collect(Collectors.toMap(objectError -> ((FieldError) objectError).getField(),
@@ -90,11 +93,12 @@ public class ExceptionHandlerEndpoint extends BaseEndpoint {
      * @return {@link ResponseDto Response body} of internal server error.
      */
     @ApiResponse(responseCode = "400", description = "The server encountered a client side error",
-            content = @Content(mediaType = APPLICATION_JSON_VALUE, schema = @Schema(implementation = ResponseDto.class)))
+            content = @Content(mediaType = APPLICATION_JSON_VALUE, schema = @Schema(implementation =
+                    ResponseDto.class)))
     @ExceptionHandler(RestException.class)
     @ResponseStatus(HttpStatus.BAD_REQUEST)
     public ResponseDto<Void> handleRuntimeException(RestException ex) {
-        logger.error("HTTP message error! " + ex.getGeneralStatus(), ex);
+        logger.error("Encountered a client side error! {}", kv("status", ex.getGeneralStatus()), ex);
         return getResponseBuilder(Void.class)
                 .setGeneralStatus(ex.getGeneralStatus())
                 .setArgs(ex.getArgs())
@@ -108,11 +112,12 @@ public class ExceptionHandlerEndpoint extends BaseEndpoint {
      * @return {@link ResponseDto Response body} of internal server error.
      */
     @ApiResponse(responseCode = "500", description = "The server encountered an internal error",
-            content = @Content(mediaType = APPLICATION_JSON_VALUE, schema = @Schema(implementation = ResponseDto.class)))
+            content = @Content(mediaType = APPLICATION_JSON_VALUE, schema = @Schema(implementation =
+                    ResponseDto.class)))
     @ExceptionHandler(RuntimeException.class)
     @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
     public ResponseDto<Void> handleRuntimeException(RuntimeException ex) {
-        logger.error("HTTP message error! " + GeneralStatus.INTERNAL_SERVER_ERROR, ex);
+        logger.error("Encountered an internal error! {}", kv("status", GeneralStatus.INTERNAL_SERVER_ERROR), ex);
         return getResponseBuilder(Void.class)
                 .setGeneralStatus(GeneralStatus.INTERNAL_SERVER_ERROR)
                 .build();
